Add files via upload
Online Rental Book System in Java

This project is an online rental book system implemented in Java, designed to facilitate the borrowing and management of books by users. The system allows users to browse a catalog of available books, rent them for a specified period, and return them once finished. It also includes features such as user authentication, book reservations, and administrative functionalities for managing books and users.